Add support for webpack and get rid of THREE.js CDN
Get rid of the need for CDN by finally adding a bundler. Use "npm run build" to run the webpack command, still "npm start" for the express server.

In doing so, I changed the general file structure of the project. It is necessary that all the files that need to be targeted for webpack be segregated from the rest (currently the two files in /src folder). Everything needed by the browser is in the /public folder. That is also the target folder for the webpack bundler.

--Note: there was some weird behavior with CSS when using webpack where the "style.fill" functionality needed to use the "gitRave()" command doesn't work. I will put in an issue for that.