Quality of life
Having server.js also console.log the current ip on working on. Convenient for development. Also tried to make it only work if it's being hosted locally and not on heroku, although i'm not sure if this is going to work (it should). There is definitely a better way to do this but for now this will do i guess.